Reverting the move of the operator implementation. Don't want any unintended consequences.
